empezando con tractor, se separan layers para renderizacion mas optima

+++ generator/v3/config/resolve_short_config.py
@@ -158,7 +158,7 @@ def resolve_short_config(
         "format": {
             "code": format_code,
             "display_name": fmt.get("display_name"),
-            "type": fmt.get("format", "short"),
+            "type": fmt.get("format", "long"),
             "engine": fmt.get("engine"),
         },
         "content": content,
